* Cool VL Viewer v1.26.1.3 (experimental, MESH branch) - The "incompletely rezzing mesh objects" bug has been fixed. As a result, the "Load Meshes at Max LOD" setting (that was used as a work around) is now off by default. - Added the prims cost (equivalent number of prims for mesh objects) to the build floater. Whenever the prim cost differs from the total number of actual prims in an object, it is displayed after that number, between parenthesis. - Added the textured light settings to the Features tab of the build floater. - Various minor improvements, backported from v3, and various minor optimizations by me.

+++ indra/newview/llselectmgr.cpp
@@ -785,7 +785,7 @@ LLObjectSelectionHandle LLSelectMgr::setHoverObject(LLViewerObject *objectp, S32
 	}
 
 	// Can't select yourself
-	if (objectp->mID == gAgentID)
+	if (objectp->mID == gAgentID && !LLSelectMgr::getInstance()->mAllowSelectAvatar)
 	{
 		mHoverObjects->deleteAllNodes();
 		return NULL;
